% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillpart.R
\name{hillpart}
\alias{hillpart}
\title{Hill numbers diversity partitioning}
\usage{
hillpart(data, q = c(0, 1, 2), tree, dist, tau)
}
\arguments{
\item{data}{A vector or a matrix/data.frame indicating the (relative) counts of one or multiple samples, respectively. If a matrix/data.frame is provided, columns must refer to samples and rows to OTUs.}

\item{q}{A positive integer or decimal number (>=0), or a vector of numbers, usually between 0 and 3.}

\item{tree}{A phylogenetic tree of class 'phylo'. The tip labels must match the names of the vector values (if one sample) or matrix rows (if multiple samples). Use the function match_data() if the OTU names do not match.}

\item{dist}{A distance matrix indicating the pairwise functional distances between samples.}

\item{tau}{An optional maximum distance value between OTUs/ASVs/MAGs to be considered in the functional Hill numbers analyses.}
}
\description{
Compute diversity partitioning of neutral, phylogenetic or functional Hill numbers from a set of samples.
}
\details{
Hill numbers diversity partitioning
}
\examples{
hillpart(data=counts)
hillpart(data=counts,tree=tree)
hillpart(data=counts,dist=dist)
hillpart(data=counts,q=0)
hillpart(data=counts,q=1,tree=tree)
hillpart(data=counts,q=2,dist=dist)
}
\references{
Jost, L. (2006). Entropy and diversity. Oikos, 113, 363-375.\cr\cr
Hill, M. O. (1973). Diversity and evenness: a unifying notation and its consequences. Ecology, 54, 427-432.\cr\cr
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.
}
\seealso{
\code{\link{hilldiv}}, \code{\link{hillpair}}, \code{\link{hilldiss}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{ASV}
\keyword{MAG}
\keyword{OTU}
\keyword{diversity}
\keyword{partitioning}
