% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillsim.R
\name{hillsim}
\alias{hillsim}
\title{Hill number similarity computation}
\usage{
hillsim(data, q = c(0, 1, 2), metric = c("S", "C", "U", "V"), tree, dist, tau)
}
\arguments{
\item{data}{A matrix/data.frame indicating the (relative) OTU/ASV/MAG counts of multiple samples. Columns must refer to samples and rows to OTUs/ASVs/MAGs.}

\item{q}{A positive integer or decimal number (>=0), or a vector of numbers, usually between 0 and 3.}

\item{metric}{Similarity metric(s) to be computed. Default is all four: c("S","C","U","V")}

\item{tree}{A phylogenetic tree of class 'phylo'. The tip labels must match the names of the vector values (if one sample) or matrix rows (if multiple samples). Use the function match_data() if the OTU names do not match.}

\item{dist}{A distance matrix indicating the pairwise functional distances between samples.}

\item{tau}{An optional maximum distance value between OTUs/ASVs/MAGs to be considered in the functional Hill numbers analyses.}
}
\description{
Compute similarity metrics based on neutral, phylogenetic and/or functional Hill numbers.
}
\details{
Hill numbers-based similarity computation
}
\examples{
hillsim(data=counts)
hillsim(data=counts,tree=tree)
hillsim(data=counts,dist=dist)
hillsim(data=counts,q=0)
hillsim(data=counts,q=1,tree=tree)
hillsim(data=counts,q=2,dist=dist)
}
\references{
Jost, L. (2006). Entropy and diversity. Oikos, 113, 363-375.\cr\cr
Hill, M. O. (1973). Diversity and evenness: a unifying notation and its consequences. Ecology, 54, 427-432.\cr\cr
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.
}
\seealso{
\code{\link{hilldiv}}, \code{\link{hillpair}}, \code{\link{hillpart}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{ASV}
\keyword{MAG}
\keyword{OTU}
\keyword{dissimilarity}
\keyword{diversity}
\keyword{overlap}
\keyword{similarity}
