% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2dist.R
\name{list2dist}
\alias{list2dist}
\title{List to dist}
\usage{
list2dist(distlist)
}
\arguments{
\item{distlist}{A table containing pairwise distances between samples, with sample names in the first 2 columns and the distance in the third column,}
}
\description{
Convert a long table of pairwise distances to a distance matrix
}
\details{
Convert a long table of pairwise distances to a distance matrix
}
\examples{
list2dist(distlist)
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{ASV}
\keyword{MAG}
\keyword{OTU}
\keyword{diversity}
