% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits2dist.R
\name{traits2dist}
\alias{traits2dist}
\title{Hill numbers computation}
\usage{
traits2dist(traits, method = "gower")
}
\arguments{
\item{traits}{A table containing trait information for each OTU/ASV/MAG in columns and OTUs/ASVs/MAGs in rows.}

\item{method}{The distance metric to be applied: "euclidean", "manhattan" or "gower" (default: "gower").}
}
\description{
Compute neutral, phylogenetic and/or functional Hill numbers from a single sample (vector) or count table (matrix).
}
\details{
Hill numbers computation
}
\examples{
traits2dist(trait,method="gower")
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{ASV}
\keyword{MAG}
\keyword{OTU}
\keyword{diversity}
